project/requirements.txt
streamlit
PyPDF2
langchain
google-generativeai
faiss-cpu
